+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - On iOS, add support for deferring system gestures.
 - On iOS, fix a crash that occurred while acquiring a monitor's name.
 - On iOS, fix armv7-apple-ios compile target.
+- Removed the `T: Clone` requirement from the `Clone` impl of `EventLoopProxy<T>`.
 
 # 0.20.0 Alpha 2 (2019-07-09)